+++ src/OpenTelemetry.Instrumentation.AspNet/CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## Unreleased
 
-* Added InstrumentationFilter public API on AspNetInstrumentationOptions to allow
-  filtering on instrumentation based on HttpContext.
+* Added Filter public API on AspNetInstrumentationOptions to allow
+  filtering of instrumentation based on HttpContext.
 
 * Asp.Net Instrumentation automatically populates HttpRequest, HttpResponse in
   Activity custom property